+++ src/components/common/CornerPattern.tsx
